+++ addOns/automation/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for data jobs.
 - Support for multiple top level URLs in a context.
 - Passive scan config enableTags parameter
+- Support for include/exclude regexes.
 
 ### Changed
 - Maintenance changes.